backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.30.0
google-genai>=1.0.0
